Update Icons path and fix console warning about key
