--- public/assets/svg/Icons.tsx
@@ -1,39 +1,0 @@
-export const MoonIcon = ({ theme }: { theme: string }) => {
-  return (
-    <svg
-      width="15"
-      height="15"
-      viewBox="0 0 24 24"
-      fill={theme === "dark" ? "black" : "white"}
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M21 12.79A9 9 0 1 1 11.21 3 7 7 0 0 0 21 12.79z"
-        stroke={theme === "dark" ? "white" : "black"}
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  );
-};
-
-export const SunIcon = ({ theme }: { theme: string }) => {
-  return (
-    <svg
-      width="15"
-      height="15"
-      viewBox="0 0 24 24"
-      fill={theme === "dark" ? "white" : "black"}
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <circle cx="12" cy="12" r="4" stroke="white" strokeWidth="2" />
-      <path
-        d="M12 2v2M12 20v2M4.93 4.93l1.41 1.41M17.66 17.66l1.41 1.41M2 12h2M20 12h2M6.34 17.66l-1.41 1.41M19.07 4.93l-1.41 1.41"
-        stroke={theme === "dark" ? "white" : "black"}
-        strokeWidth="2"
-        strokeLinecap="round"
-      />
-    </svg>
-  );
-};
