+++ src/App.module.css
@@ -7,5 +7,5 @@
   background-color: var(--color-background-white);
   height: 100dvh;
   transition: background-color 0.2s ease-in-out, color 0.2s ease-in-out;
-  border: 1px dashed var(--color-background-black);
+  /* border: 1px solid transparent; */
 }
