+++ src/components/MidSection/Languages/Languages.module.css
@@ -7,6 +7,7 @@
   align-items: center;
   gap: 3rem;
   background-color: var(--color-background-dark);
+  border-top: 1px solid var(--color-background-white);
 }
 
 .languages__css {
@@ -41,7 +42,7 @@
   color: var(--color-text-black);
   -webkit-text-stroke: 1px var(--color-background-html);
   border: none;
-  box-shadow: 0 0 4px 3px #404040;
+  box-shadow: 0 0 2px 2px #404040;
   transition: all 0.3s ease-in-out;
   transition: border-radius 0.1s ease-in-out;
   &:hover {
